[Refactor] #129 — 패키지 위치 이동
