--- src/main/java/com/yanus/attendance/attendance/domain/AttendanceStatus.java
@@ -1,5 +1,0 @@
-package com.yanus.attendance.attendance.domain;
-
-public enum AttendanceStatus {
-    WORKING, LEFT
-}
