+++ src/main/java/com/yanus/attendance/attendance/domain/attendance/AttendanceStatus.java
@@ -1,0 +1,5 @@
+package com.yanus.attendance.attendance.domain.attendance;
+
+public enum AttendanceStatus {
+    WORKING, LEFT
+}
