+++ src/main/java/com/yanus/attendance/attendance/domain/workschedule/WeekPattern.java
@@ -1,4 +1,4 @@
-package com.yanus.attendance.attendance.domain;
+package com.yanus.attendance.attendance.domain.workschedule;
 
 public enum WeekPattern {
     EVERY, FIRST, SECOND, THIRD, FOURTH, LAST
