+++ src/main/java/com/yanus/attendance/attendance/infrastructure/AttendanceQueryRepositoryImpl.java
@@ -2,8 +2,8 @@ package com.yanus.attendance.attendance.infrastructure;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.yanus.attendance.attendance.domain.Attendance;
-import com.yanus.attendance.attendance.domain.AttendanceQueryRepository;
+import com.yanus.attendance.attendance.domain.attendance.Attendance;
+import com.yanus.attendance.attendance.domain.attendance.AttendanceQueryRepository;
 import com.yanus.attendance.attendance.domain.QAttendance;
 import com.yanus.attendance.member.domain.QMember;
 import com.yanus.attendance.team.domain.QTeam;
