+++ src/main/java/com/yanus/attendance/attendance/infrastructure/WorkScheduleJpaRepository.java
@@ -1,7 +1,7 @@
 package com.yanus.attendance.attendance.infrastructure;
 
-import com.yanus.attendance.attendance.domain.WorkSchedule;
-import com.yanus.attendance.attendance.domain.WorkScheduleRepository;
+import com.yanus.attendance.attendance.domain.workschedule.WorkSchedule;
+import com.yanus.attendance.attendance.domain.workschedule.WorkScheduleRepository;
 import java.time.DayOfWeek;
 import java.util.List;
 import java.util.Optional;
