+++ src/main/java/com/yanus/attendance/attendance/presentation/dto/WorkScheduleRequest.java
@@ -1,6 +1,6 @@
 package com.yanus.attendance.attendance.presentation.dto;
 
-import com.yanus.attendance.attendance.domain.WeekPattern;
+import com.yanus.attendance.attendance.domain.workschedule.WeekPattern;
 import java.time.DayOfWeek;
 import java.time.LocalTime;
