+++ src/main/java/com/yanus/attendance/attendance/presentation/dto/WorkScheduleResponse.java
@@ -1,7 +1,7 @@
 package com.yanus.attendance.attendance.presentation.dto;
 
-import com.yanus.attendance.attendance.domain.WeekPattern;
-import com.yanus.attendance.attendance.domain.WorkSchedule;
+import com.yanus.attendance.attendance.domain.workschedule.WeekPattern;
+import com.yanus.attendance.attendance.domain.workschedule.WorkSchedule;
 import java.time.DayOfWeek;
 import java.time.LocalTime;
